+++ vite.config.ts
@@ -20,6 +20,7 @@ export default defineConfig({
             "@xterm/addon-fit",
             "@xterm/addon-web-links",
             "@xterm/addon-unicode11",
+            "@xterm/addon-webgl",
           ],
         },
       },